Use QlineEdit for name and version
- Make sure version is limited to number and from 1-100.

+++ core_lib/src/util/mpbrushutils.h
@@ -507,7 +507,7 @@ struct MPCONF {
 
 struct MPBrushInfo {
     QString comment = "";
-    qreal version = 0.0;
+    int version = 0;
 
     void write(QJsonObject& object) const
     {
@@ -543,7 +543,7 @@ struct MPBrushInfo {
         QString brushVersionKey = "brush-version";
         QJsonObject::iterator versionObjIt = object.find(brushVersionKey);
         if (!versionObjIt->isUndefined()) {
-            info.version = object.value(brushVersionKey).toDouble();
+            info.version = object.value(brushVersionKey).toInt();
         }
 
         return info;
